requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiofiles==24.1.0
    # via sanic
async-timeout==4.0.3
    # via redis
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
biotech==1.1.10
    # via
    #   law-dictionary
    #   ships
    #   somatic
    #   ventures
blinker==1.8.2
    # via flask
body-scan==1.1.1
    # via
    #   biotech
    #   emergency
botanist==1.0.0
    # via
    #   biotech
    #   body-scan
    #   emergency
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   foam-pet (pyproject.toml)
    #   biotech
    #   body-scan
    #   emergency
    #   flask
    #   law-dictionary
    #   somatic
    #   ventures
coverage==7.6.2
    # via
    #   biotech
    #   emergency
cx-freeze==7.2.3
    # via foam-pet (pyproject.toml)
dnspython==2.7.0
    # via pymongo
emergency==2.0.0
    # via foam-pet (pyproject.toml)
filelock==3.16.1
    # via cx-freeze
flask==3.0.3
    # via
    #   biotech
    #   body-scan
    #   emergency
    #   somatic
html5tagger==1.3.0
    # via
    #   sanic
    #   tracerite
httptools==0.6.1
    # via sanic
idna==3.10
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
jsonschema==4.23.0
    # via somatic
jsonschema-specifications==2024.10.1
    # via jsonschema
law-dictionary==1.0.12
    # via somatic
mako==1.3.5
    # via
    #   biotech
    #   emergency
    #   somatic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.1
    # via
    #   jinja2
    #   mako
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
multidict==6.1.0
    # via sanic
packaging==24.1
    # via
    #   cx-freeze
    #   pipdeptree
patchelf==0.17.2.1
    # via cx-freeze
pexpect==4.9.0
    # via
    #   biotech
    #   emergency
pip==24.2
    # via pipdeptree
pip-licenses==5.0.0
    # via foam-pet (pyproject.toml)
pipdeptree==2.23.4
    # via foam-pet (pyproject.toml)
prettytable==3.11.0
    # via pip-licenses
psutil==5.9.8
    # via
    #   botanist
    #   ships
ptyprocess==0.7.0
    # via pexpect
pydash==8.0.3
    # via
    #   foam-pet (pyproject.toml)
    #   ships
    #   ventures
pygments==2.18.0
    # via
    #   emergency
    #   rich
pyinotify==0.9.6
    # via ships
pymongo==4.10.1
    # via
    #   biotech
    #   emergency
pyyaml==6.0.2
    # via sanic-ext
redis==5.1.1
    # via
    #   biotech
    #   emergency
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   foam-pet (pyproject.toml)
    #   biotech
    #   body-scan
    #   emergency
rich==13.9.2
    # via
    #   foam-pet (pyproject.toml)
    #   biotech
    #   emergency
    #   ships
    #   somatic
    #   ventures
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
sanic==23.12.2
    # via
    #   foam-pet (pyproject.toml)
    #   biotech
    #   emergency
    #   somatic
    #   ventures
sanic-ext==23.12.0
    # via ventures
sanic-routing==23.12.0
    # via sanic
setuptools==69.5.1
    # via
    #   cx-freeze
    #   ventures
ships==1.9.0
    # via foam-pet (pyproject.toml)
somatic==3.0.3
    # via
    #   foam-pet (pyproject.toml)
    #   ventures
tinydb==4.8.2
    # via
    #   biotech
    #   body-scan
    #   emergency
tomli==2.0.2
    # via pip-licenses
tracerite==1.1.1
    # via sanic
typing-extensions==4.12.2
    # via
    #   pydash
    #   sanic
ujson==5.10.0
    # via sanic
urllib3==2.2.3
    # via requests
uvloop==0.20.0
    # via sanic
ventures==3.0.0
    # via foam-pet (pyproject.toml)
wcwidth==0.2.13
    # via prettytable
websockets==13.1
    # via sanic
werkzeug==3.0.4
    # via flask
